Работает добавление и удаление картинки с фронта

+++ src/main/resources/liquibase/scripts/main.sql
@@ -29,9 +29,9 @@ CREATE TABLE comments (
 
 -- changeSet yuvis:4
 CREATE TABLE pictures (
-    id SERIAL NOT NULL PRIMARY KEY,
+    id BIGINT NOT NULL PRIMARY KEY,
     file_path TEXT NULL,
-    file_size INTEGER NULL,
+    file_size BIGINT NULL,
     media_type VARCHAR(255),
     data TEXT
 );
